SWATCH-1880: Move purge Events from swatch-metrics to swatch-tally
