+++ src/main/java/org/candlepin/subscriptions/tally/events/EventRecordsRetentionProperties.java
@@ -18,7 +18,7 @@
  * granted to use or replicate Red Hat trademarks that are incorporated
  * in this software or its documentation.
  */
-package org.candlepin.subscriptions.metering.retention;
+package org.candlepin.subscriptions.tally.events;
 
 import java.time.Duration;
 import lombok.Getter;
