+++ src/main/spec/internal-tally-api-spec.yaml
@@ -233,6 +233,18 @@ paths:
         '500':
           $ref: "../../../spec/error-responses.yaml#/$defs/InternalServerError"
       tags: [ internalTally ]
+  /internal/rpc/tally/events/purge:
+    description: 'Purge event records'
+    delete:
+      operationId: purgeEventRecords
+      responses:
+        '200':
+          description: "Purge was successful."
+        '403':
+          $ref: "../../../spec/error-responses.yaml#/$defs/Forbidden"
+        '500':
+          $ref: "../../../spec/error-responses.yaml#/$defs/InternalServerError"
+      tags: [internalTally]
   /internal/rpc/tally/events/{event_id}:
     description: Delete an event. Supported only in dev-mode.
     delete:
