+++ src/test/java/org/candlepin/subscriptions/tally/admin/InternalTallyResourceTest.java
@@ -30,11 +30,13 @@ import jakarta.ws.rs.BadRequestException;
 import java.time.OffsetDateTime;
 import org.candlepin.clock.ApplicationClock;
 import org.candlepin.subscriptions.ApplicationProperties;
+import org.candlepin.subscriptions.db.EventRecordRepository;
 import org.candlepin.subscriptions.retention.RemittanceRetentionController;
 import org.candlepin.subscriptions.retention.TallyRetentionController;
 import org.candlepin.subscriptions.security.SecurityProperties;
 import org.candlepin.subscriptions.tally.MarketplaceResendTallyController;
 import org.candlepin.subscriptions.tally.TallySnapshotController;
+import org.candlepin.subscriptions.tally.events.EventRecordsRetentionProperties;
 import org.candlepin.subscriptions.tally.job.CaptureSnapshotsTaskManager;
 import org.candlepin.subscriptions.test.TestClockConfiguration;
 import org.candlepin.subscriptions.util.DateRange;
@@ -54,6 +56,8 @@ class InternalTallyResourceTest {
   @Mock private RemittanceRetentionController remittanceRetentionController;
   @Mock private InternalTallyDataController internalTallyDataController;
   @Mock private SecurityProperties properties;
+  @Mock private EventRecordRepository eventRecordRepository;
+  @Mock private EventRecordsRetentionProperties eventRecordsRetentionProperties;
 
   private InternalTallyResource resource;
   private ApplicationProperties appProps;
@@ -73,7 +77,9 @@ class InternalTallyResourceTest {
             tallyRetentionController,
             remittanceRetentionController,
             internalTallyDataController,
-            properties);
+            properties,
+            eventRecordRepository,
+            eventRecordsRetentionProperties);
   }
 
   @Test
